requirements.txt
# ===========================================
# Web Scraper Application - Dependencies
# ===========================================
# Web Scraping Tool for TikTok & Facebook
# Python 3.9+
# Last Updated: January 2026

# ===========================================
# Core Framework
# ===========================================
Flask==3.0.0                    # Web framework
Flask-Login==0.6.3              # User session management
Flask-WTF==1.2.1                # CSRF protection & forms
Flask-SQLAlchemy==3.1.1         # ORM integration
WTForms==3.1.1                  # Form rendering & validation

# ===========================================
# Database
# ===========================================
SQLAlchemy==2.0.23              # SQL toolkit and ORM
psycopg2-binary==2.9.9          # PostgreSQL adapter for Python

# ===========================================
# Security & Authentication
# ===========================================
Werkzeug==3.0.1                 # Password hashing (PBKDF2)
python-dotenv==1.0.0            # Environment variables
email-validator==2.1.0          # Email validation
Flask-Mail==0.10.0              # Email sending for password reset
itsdangerous==2.2.0             # Token generation

# ===========================================
# Web Scraping
# ===========================================
selenium==4.15.2                # Browser automation
webdriver-manager==4.0.1        # Automatic ChromeDriver management

# ===========================================
# Data Export
# ===========================================
pandas==2.2.0                   # Data manipulation
openpyxl==3.1.2                 # Excel file creation

# ===========================================
# Utilities
# ===========================================
requests==2.31.0                # HTTP library

# ===========================================
# Production Server
# ===========================================
gunicorn==21.2.0                # WSGI HTTP Server
